Move all `resources_sender` functions into `peer_handler`
They aren't used anywhere else. Thus, they can all be private to this module.
